+++ src/contexts/index.ts
@@ -1,7 +1,5 @@
-export {
-  GroupsProvider,
-  isUnauthorizedError,
-  useGroupActions,
-  useGroups,
-} from './groups-context'
-export type { GroupActions, GroupsData, RecentGroup } from './groups-context'
+export { GroupsProvider } from './groups/core-context'
+export { isUnauthorizedError } from './groups/helpers'
+export type { GroupActions, GroupsData, RecentGroup } from './groups/types'
+export { useGroupActions } from './groups/use-group-actions'
+export { useGroups } from './groups/use-groups'
